LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/content-security-policy/reporting-api/reporting-api-report-only-sends-reports-on-violation.https.sub-expected.txt



PASS Test that image does not load
PASS Event is fired
FAIL Violation report status OK. undefined is not an object (evaluating 'data[0]["body"]')
